+++ varnish-sys/src/vcl/convert.rs
@@ -52,7 +52,7 @@ use std::ptr::{null, null_mut};
 use std::time::{Duration, SystemTime};
 
 use crate::ffi::{
-    sa_family_t, vsa_suckaddr_len, vtim_dur, vtim_real, VSA_BuildFAP, VSA_GetPtr, VSA_Port,
+    sa_family_t, vsa_suckaddr_len, vtim_dur, vtim_real, /*VSA_BuildFAP, */VSA_GetPtr, VSA_Port,
     PF_INET, PF_INET6, VCL_ACL, VCL_BACKEND, VCL_BLOB, VCL_BODY, VCL_BOOL, VCL_DURATION, VCL_ENUM,
     VCL_HEADER, VCL_HTTP, VCL_INT, VCL_IP, VCL_PROBE, VCL_REAL, VCL_REGEX, VCL_STEVEDORE,
     VCL_STRANDS, VCL_STRING, VCL_SUB, VCL_TIME, VCL_VCL,
@@ -188,6 +188,7 @@ impl From<VCL_INT> for i64 {
 //
 // VCL_IP
 //
+/*
 default_null_ptr!(VCL_IP);
 impl IntoVCL<VCL_IP> for SocketAddr {
     fn into_vcl(self, ws: &mut Workspace) -> Result<VCL_IP, VclError> {
@@ -227,6 +228,7 @@ impl IntoVCL<VCL_IP> for SocketAddr {
         }
     }
 }
+*/
 impl From<VCL_IP> for Option<SocketAddr> {
     fn from(value: VCL_IP) -> Self {
         let value = value.0;
