+++ varnish-sys/src/vcl/ctx.rs
@@ -86,7 +86,7 @@ impl<'a> Ctx<'a> {
             }
         }
     }
-
+/*
     pub fn cached_req_body(&mut self) -> Result<Vec<&'a [u8]>, VclError> {
         unsafe extern "C" fn chunk_collector(
             priv_: *mut c_void,
@@ -121,6 +121,7 @@ impl<'a> Ctx<'a> {
             _ => Err("req.body iteration failed".into()),
         }
     }
+*/
 }
 
 /// A struct holding both a native [`vrt_ctx`] struct and the space it points to.
@@ -157,7 +158,7 @@ pub fn log(tag: LogTag, msg: impl AsRef<str>) {
     unsafe {
         ffi::VSL(
             tag,
-            ffi::vxids { vxid: 0 },
+            0, //ffi::vxids { vxid: 0 },
             c"%.*s".as_ptr(),
             msg.len(),
             msg.as_ptr(),
@@ -173,5 +174,32 @@ mod tests {
     fn ctx_test() {
         let mut test_ctx = TestCtx::new(100);
         test_ctx.ctx();
+    }
+}
+
+/// This is an unsafe struct that holds the per-VCL state.
+/// It must be public because it is used by the macro-generated code.
+#[doc(hidden)]
+#[derive(Debug)]
+pub struct PerVclState<T> {
+//    pub fetch_filters: Vec<Box<ffi::vfp>>,
+//    pub delivery_filters: Vec<Box<ffi::vdp>>,
+    pub user_data: Option<Box<T>>,
+}
+
+// Implement the default trait that works even when `T` does not impl `Default`.
+impl<T> Default for PerVclState<T> {
+    fn default() -> Self {
+        Self {
+//            fetch_filters: Vec::default(),
+//            delivery_filters: Vec::default(),
+            user_data: None,
+        }
+    }
+}
+
+impl<T> PerVclState<T> {
+    pub fn get_user_data(&self) -> Option<&T> {
+        self.user_data.as_ref().map(AsRef::as_ref)
     }
 }
