+++ varnish-sys/src/vcl/mod.rs
@@ -1,20 +1,20 @@
-mod backend;
+//mod backend;
 mod convert;
 mod ctx;
 mod error;
 mod http;
 mod probe;
-mod processor;
+//mod processor;
 mod vsb;
 mod ws;
 
-pub use backend::*;
+//pub use backend::*;
 pub use convert::*;
 pub use ctx::*;
 pub use error::*;
 pub use http::*;
 pub use probe::*;
-pub use processor::*;
+//pub use processor::*;
 pub use vsb::*;
 pub use ws::*;
